feat: refactor guides structure and layout
